pyproject.toml

[build-system]
requires = ["setuptools-scm[toml]>=7", "setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "qrSVG"
description = "QR Code SVG generator"
dynamic = ["version"]

requires-python = ">=3.12"
dependencies = [ #
    "CairoSVG",
    "inquirer",
    "numpy",
    "opencv-python",
    "pillow",
    "qrcode",
]

[project.optional-dependencies]
dev = [ #
    "mypy",
    "pre-commit",
    "pytest",
    "ruff",
    "textual-dev",
    "tox",
]

[project.scripts]
qrsvg = "qrSVG.cli:main"
qrsvg-meta = "qrSVG.diff:metadata"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
namespaces = false
where = ["src"]


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [ #
    "B",
    "C",
    "E",
    "F",
    "FLY",
    "G",
    "I",
    "PL",
    "Q",
    "RUF",
    "UP",
    "W",
]

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
